requirements.txt
python-dotenv
pymupdf
pdfplumber
langchain
chromadb
faiss-cpu
sentence-transformers
openai
tiktoken
pandas
numpy
matplotlib
plotly
yfinance
prophet
requests
beautifulsoup4
newsapi-python
transformers
accelerate
gradio
streamlit
langchain-tavily 
langchain-google-genai
ipython
gradio